+++ apps/api/src/main/java/br/org/apae/api/professional/exceptions/InvalidDataException.java
@@ -1,4 +1,4 @@
-package br.org.apae.api.professional.domain.exceptions;
+package br.org.apae.api.professional.exceptions;
 
 public class InvalidDataException extends RuntimeException {
